Removed lots of files. Now running off main.cpp.
