+++ include/core0.h
@@ -1,139 +1,34 @@
-#ifndef Arduino
-#include <Arduino.h>
-#endif
 
-#include <RF24Network.h>
-#include <NRF24L01Lib.h>
-#include <TriggerLib.h>
-
-#include <comms_fsm.h>
-
-//------------------------------------------------------------
-RF24 radio(SPI_CE, SPI_CS);
-RF24Network network(radio);
-NRF24L01Lib nrf24;
-
-void board_packet_available_cb(uint16_t from_id, uint8_t type);
-void send_controller_packet_to_board();
-
-#define SEND_TO_BOARD_MS  3000
-#define NUM_RETRIES       2
-#define LOG_LENGTH_MILLIS 5000
-
-Smoothed <float> retry_log;
-
-void comms_task_0(void *pvParameters)
+void display_task_0(void *pvParameters)
 {
-  comms_fsm.run_machine();
+  elapsedMillis since_show_retry_log;
 
-  nrf24.begin(&radio, &network, /*address*/ 1, board_packet_available_cb);
+  setupLCD();
 
-  Serial.printf("comms_task_0 running on core %d\n", xPortGetCoreID());
+  Serial.printf("display_task_0 running on core %d\n", xPortGetCoreID());
 
-  retry_log.begin(SMOOTHED_AVERAGE, LOG_LENGTH_MILLIS / SEND_TO_BOARD_MS);
-
-  add_comms_fsm_transitions();
-
-  elapsedMillis sent_to_board;
+  int i = 0;
 
   while (true)
   {
-    comms_fsm.run_machine();
-
-    if (since_(sent_to_board, SEND_TO_BOARD_MS, false))
+    if (since_show_retry_log > 2000)
     {
-      if (since_(sent_to_board, SEND_TO_BOARD_MS + 10, false))
-      {
-        DEBUGVAL(sent_to_board);
-      }
-      sent_to_board = 0;
+      since_show_retry_log = 0;
 
-      send_controller_packet_to_board();
+      char buff1[20];
+      char buff2[20];
+      u8g2.clearBuffer();
+      // line 1
+      sprintf(buff1, "rate: %.1f%%", retry_log.get());
+      lcd_message(1, buff1, Aligned::ALIGNED_LEFT);
+      // line 2
+      sprintf(buff2, "total: %lu", stats.total_failed);
+      lcd_message(2, buff2, Aligned::ALIGNED_LEFT);
+      // line 3
+      // lcd_message(3, buff, Aligned::ALIGNED_LEFT);
+      u8g2.sendBuffer();
     }
-
-    nrf24.update();
-
     vTaskDelay(10);
-  }
-  vTaskDelete(NULL);
-}
-//-------------------------------------------------
-void send_controller_packet_to_board()
-{
-  controller_packet.id++;
-  uint8_t bs[sizeof(ControllerData)];
-  memcpy(bs, &controller_packet, sizeof(ControllerData));
-
-  DEBUGVAL("send_controller_packet_to_board");
-  uint8_t retries = nrf24.send_with_retries(00, /*type*/0, bs, sizeof(ControllerData), NUM_RETRIES);
-  DEBUGVAL(retries);
-  retry_log.add(retries > 0);
-
-  if (retries > 0)
-  {
-    DEBUGVAL(retries, retry_log.get());
-  }
-
-  controller_packet.command = 0;
-}
-//-------------------------------------------------
-void board_packet_available_cb(uint16_t from_id, uint8_t type)
-{
-  uint8_t buff[sizeof(VescData)];
-  nrf24.read_into(buff, sizeof(VescData));
-  memcpy(&board_packet, &buff, sizeof(VescData));
-  DEBUGVAL(board_packet.id);
-
-  switch (board_packet.reason)
-  {
-    case ReasonType::BOARD_STOPPED:
-      DEBUG("BOARD_STOPPED");
-      break;
-    case ReasonType::BOARD_MOVING:
-      DEBUG("BOARD_MOVING");
-      break;
-  }
-
-  // DEBUGVAL(board_packet.id);
-}
-//------------------------------------------------------------
-
-void trigger_read_task_0(void *pvParameters)
-{
-  elapsedMillis since_read_trigger;
-  uint8_t old_throttle = 0;
-
-  #define READ_TRIGGER_PERIOD 100
-  #define SMOOTH_OVER_MILLIS  2000
-
-  TriggerLib trigger(10);
-  trigger.initialise();
-  
-  Smoothed <int> sm_throttle;
-  sm_throttle.begin(SMOOTHED_EXPONENTIAL, SMOOTH_OVER_MILLIS / READ_TRIGGER_PERIOD);
-
-  Serial.printf("\trigger_read_task_0 running on core %d\n", xPortGetCoreID());
-
-  DEBUGVAL("num smooths", SMOOTH_OVER_MILLIS / READ_TRIGGER_PERIOD);
-
-  while (true)
-  {
-    if (since_read_trigger > READ_TRIGGER_PERIOD)
-    {
-      since_read_trigger = 0;
-
-      uint8_t throttle = trigger.get_throttle();
-      sm_throttle.add(throttle);
-      controller_packet.throttle = sm_throttle.get();
-
-      if (old_throttle != controller_packet.throttle)
-      {
-        DEBUGVAL(controller_packet.throttle);
-        old_throttle = controller_packet.throttle;
-      }
-    }
-
-    vTaskDelay(1);
   }
   vTaskDelete(NULL);
 }
