--- include/display_0.h
@@ -1,34 +1,0 @@
-
-void display_task_0(void *pvParameters)
-{
-  elapsedMillis since_show_retry_log;
-
-  setupLCD();
-
-  Serial.printf("display_task_0 running on core %d\n", xPortGetCoreID());
-
-  int i = 0;
-
-  while (true)
-  {
-    if (since_show_retry_log > 2000)
-    {
-      since_show_retry_log = 0;
-
-      char buff1[20];
-      char buff2[20];
-      u8g2.clearBuffer();
-      // line 1
-      sprintf(buff1, "rate: %.1f%%", retry_log.get());
-      lcd_message(1, buff1, Aligned::ALIGNED_LEFT);
-      // line 2
-      sprintf(buff2, "total: %lu", stats.total_failed);
-      lcd_message(2, buff2, Aligned::ALIGNED_LEFT);
-      // line 3
-      // lcd_message(3, buff, Aligned::ALIGNED_LEFT);
-      u8g2.sendBuffer();
-    }
-    vTaskDelay(10);
-  }
-  vTaskDelete(NULL);
-}
