+++ platformio.ini
@@ -1,9 +1,8 @@
 [platformio]
-; default_envs = esp32dev.debug
+default_envs = esp32dev.debug
 ; default_envs = esp32dev.release
 ; default_envs = esp32dev.comms_and_display
 ; default_envs = esp32dev.comms_ping_pong
-default_envs = esp32dev.comms_with_fsm_2
 
 [env]
 platform = espressif32
@@ -19,6 +18,7 @@ lib_deps =
   https://github.com/lewisxhe/Button2.git
   https://github.com/bblanchon/ArduinoJson.git#6.x
   https://github.com/pfeerick/elapsedMillis.git
+  https://github.com/MattFryer/Smoothed.git
 
   U8g2@2.27.0
 
@@ -35,36 +35,17 @@ lib_deps =
 [env:esp32dev.debug]
 src_filter = ${dir.main}
 build_flags = 
-  -D DEBUG_BUILD
-  -D USE_TEST_VALUES
-  -D PRINT_FSM_EVENT
-  ; -D USE_DEADMAN_SWITCH
-  ; -D PRINT_TRIGGER_FSM_EVENT
-  ; -D PRINT_BOARD_FSM_EVENT  
-  ; -D PRINT_BOARD_FSM_STATE_NAME     
-  ; -D PRINT_TRIGGER_VALUE 
-  -D PACKET_RECV_DEBUG_ENABLED   
-  ; -D DEBUG_PRINT_STATE_NAME_ENABLED   
-  ; -D DEBUG_PRINT_TRIGGER_STATE
-  ; -D PRINT_RETRIES_STATS
+  -D PRINT_THROTTLE
+  -D FEATURE_CRUISE_CONTROL
+  -D SERIAL__DEBUG
 ;-------------------------------------------------------
 
 [env:esp32dev.release]
 src_filter = ${dir.main}
 build_flags = 
-  -D RELEASE_BUILD
-  -D USE_TEST_VALUES
-  ; -D DEBUG_TRIGGER_ENABLED
-  ; -D DEBUG_THROTTLE_ENABLED
-  ; -D USE_DEADMAN_SWITCH
-  ; -D PRINT_TRIGGER_FSM_EVENT
-  ; -D PRINT_BOARD_FSM_STATE_NAME     
-  ; -D PRINT_BOARD_FSM_EVENT  
-  ; -D DEBUG_PRINT_STATE_NAME_ENABLED
-  ; -D PACKET_RECV_DEBUG_ENABLED   
-  ; -D USING_BUTTONS
-  -D PRINT_TRIGGER_VALUE 
-  -D PRINT_RETRIES_STATS
+  ; -D PRINT_THROTTLE
+  ; -D FEATURE_CRUISE_CONTROL
+  ; -D SERIAL__DEBUG
 ;-------------------------------------------------------
 
 [env:esp32dev.comms_and_display]
@@ -102,32 +83,8 @@ lib_deps =
   https://github.com/skelstar/NRF24L01.Library.git#feature/adding-test
 
 ;-------------------------------------------------------
-[env:esp32dev.comms_with_fsm_2]
 
-src_filter = ${dir.comms_with_fsm_2}
-
-build_flags = 
-  ; -D PRINT_THROTTLE
-  ; -D FEATURE_CRUISE_CONTROL
-
-lib_deps = 
-  https://github.com/pfeerick/elapsedMillis.git
-  https://github.com/lewisxhe/Button2.git
-  https://github.com/MattFryer/Smoothed.git
-  U8g2@2.27.0
-
-  https://github.com/skelstar/VESC_Comms.git
-  https://github.com/skelstar/Esk8.Board.VescComms.git#metadata
-  https://github.com/skelstar/Arduino-Debugging.git
-
-  https://github.com/nRF24/RF24Network.git
-  https://github.com/skelstar/NRF24L01.Library.git#feature/adding-test
-  https://github.com/skelstar/arduino-fsm.git 
-
-
-;-------------------------------------------------------
 [dir]
 main = +<main/> -<comms_only/> -<comms_and_display/>
 comms_and_display = +<comms_and_display/>
 comms_ping_pong = +<comms_ping_pong/>
-comms_with_fsm_2 = +<comms_with_fsm_2/>
